feat: fix stock deduction error

+++ posapp/models.py
@@ -3,13 +3,13 @@ from django.db import models
 
 class BahanBaku(models.Model):
     name = models.CharField(max_length=255)
-    stock = models.IntegerField(null=True)
+    stock = models.DecimalField(max_digits=5, decimal_places=2, null=True)
     satuan = models.CharField(max_length=255)
 
 class MenuItem(models.Model):
     nama_menu = models.CharField(max_length=255)
     description = models.TextField()
-    harga = models.IntegerField(null=True)
+    harga = models.DecimalField(max_digits=5, decimal_places=2, null=True)
     category = models.CharField(choices=[('beverages', 'Beverages'), ('dishes', 'Dishes'), ('dessert', 'Dessert')], max_length=10, null=True)
     bahan_baku = models.ManyToManyField(BahanBaku, null=True)
     gambar = models.ImageField(upload_to='menu_images/', null=True, blank=True)
@@ -17,12 +17,12 @@ class MenuItem(models.Model):
 class BahanBakuPerMenu(models.Model):
     menu_item = models.ForeignKey(MenuItem, on_delete=models.CASCADE, null=True)
     bahan_baku = models.ForeignKey(BahanBaku, on_delete=models.CASCADE, null=True)
-    quantity = models.PositiveIntegerField(null=True)
+    quantity = models.DecimalField(max_digits=5, decimal_places=2, null=True)
 
 class Order(models.Model):
     menu_item = models.ForeignKey(MenuItem, on_delete=models.CASCADE)
     table_number = models.IntegerField(null=True)
-    quantity = models.IntegerField(null=True)
+    quantity = models.DecimalField(max_digits=5, decimal_places=2, null=True)
     order_type = models.CharField(choices=[('online', 'Online'), ('dine-in', 'Dine-In')], max_length=10, null=True)
     alamat = models.TextField(blank=True, null=True)
     atas_nama = models.CharField(max_length=255, null=True)
